+++ test/Limiters/vertical_mass_borrowing_limiter_advection.jl
@@ -2,10 +2,10 @@ using Test
 using LinearAlgebra
 import ClimaComms
 ClimaComms.@import_required_backends
-using OrdinaryDiffEqSSPRK: ODEProblem, solve, SSPRK33
 using ClimaCore.CommonGrids
 using ClimaCore.Grids
 using ClimaTimeSteppers
+using ClimaTimeSteppers: ODEProblem, solve
 import ClimaCore
 
 import ClimaCore: